pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "beautifulsoup4>=4.13.3",
    "email-validator>=2.2.0",
    "faiss-cpu>=1.10.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.2.4",
    "openai>=1.68.2",
    "psycopg2-binary>=2.9.10",
    "pymupdf>=1.25.4",
    "pypdf2>=3.0.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.39",
    "tenacity>=9.0.0",
    "trafilatura>=2.0.0",
    "werkzeug>=3.1.3",
]
